+++ project/front/app/src/App.tsx
@@ -1,37 +1,27 @@
-import React, { useEffect, useState } from 'react';
-import './App.css';
 
-function GetTokenUser(props: { url: string }) {
-  const [body, setBody] = useState<string | null>(null);
+import React from 'react';
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Auth from './pages/Auth';
+import NotFound from './pages/NotFound';
+import Game from './pages/game';
+import Profil from './pages/profil';
+import Accueil from './pages/accueil';
 
-  useEffect(() => {
-    fetch(props.url)
-      .then((response) => response.text())
-      .then((text) => setBody(text));
-  }, []);
 
-  if (!body) {
-    return <div>Loading...</div>;
-  }
-
-  return <div dangerouslySetInnerHTML={{ __html: body }} />;
-}
 
 function App() {
-  return (
-	<div className="App">
-		<h1>Transcendence</h1>
-		<a
-		className="api42-link"
-		href="http://localhost:6200/auth/login"
-		target="_blank"
-		rel="noopener noreferrer"
-		>
-			Authentification
-		</a>
-			<GetTokenUser url="http://localhost:6200/auth/login"/>
-    </div>
-  );
+	return (
+		<BrowserRouter>
+			<Routes>
+				<Route path="/" Component={Auth}/>
+				<Route path="*" Component={NotFound}/>
+				<Route path="/accueil" Component={Accueil} />
+      			<Route path="/game" Component={Game} />
+      			<Route path="/profil" Component={Profil} />
+			</Routes>
+		</BrowserRouter>
+	);
+
 }
 
 export default App;
