+++ project/front/app/src/pages/Auth.tsx
@@ -1,4 +1,23 @@
 import React from 'react';
+import { useState, useEffect } from 'react';
+
+
+function GetTokenUser(props: { url: string }) {
+	const [body, setBody] = useState<string | null>(null);
+  
+	useEffect(() => {
+	  fetch(props.url)
+		.then((response) => response.text())
+		.then((text) => setBody(text));
+	}, []);
+  
+	if (!body) {
+	  return <div>Loading...</div>;
+	}
+  
+	return <div dangerouslySetInnerHTML={{ __html: body }} />;
+  }
+
 
 function Auth() {
   //state
